Add fields for type, medium and institution
Also, facets for type and medium are provided. Setting these facets does not
work yet, elasticsearch does not produce any results for

    https://api.ellinet-dev.hbz-nrw.de/search/frl/_search?q=medium:*

Is this the wrong query, or is this an issue in the back end?

Another known issue is that the facets are not labelled yet.

Please run update script.